requirements.txt
slack_bolt
slack_sdk
python-dotenv
requests
typing_extensions
